flask-backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
Babel==2.8.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.7.14
click==8.1.8
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.3.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jiter==0.10.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
openai==1.98.0
pydantic==2.11.7
pydantic_core==2.33.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dotenv==1.1.1
pytz==2022.1
PyYAML==5.4.1
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
systemd-python==234
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
Werkzeug==3.1.3
zipp==1.0.0
